build: Remove ORI permission system
