+++ packages/trpc-interface/src/index.server.ts
@@ -5,8 +5,8 @@ export type { AppContext } from "./context/context.server";
 
 export { AuthorizationError } from "./context/errors.server";
 export * as authorizeProject from "./authorize/project.server";
-export * as authorizeAuthorizationToken from "./authorize/authorization-token.server";
-export type { AuthPermit } from "./shared/authorization-router";
+export type { AuthPermit } from "./authorize/project.server";
+
 export {
   router,
   procedure,
