+++ packages/trpc-interface/src/shared/shared-router.ts
@@ -1,11 +1,9 @@
 import type { createTRPCProxyClient } from "@trpc/client";
 import { router } from "./trpc";
-import { authorizationRouter } from "./authorization-router";
 import { domainRouter } from "./domain";
 import { deploymentRouter } from "./deployment";
 
 export const sharedRouter = router({
-  authorize: authorizationRouter,
   domain: domainRouter,
   deployment: deploymentRouter,
 });
